Move List item content delete icon next to resource name (#60)
The reason for moving the icon to be just to the right of the resource name is so that it isn't so close to the delete icon for the entire item. Also, made it smaller to further distinguish it from the Item delete button. A horizontal line was also added to separate the header from the content area of the item, similar to the messages type.