Add delete access check for deleting interpretations and comments (#90)
According to the backend team, the delete access check is as follows:

To delete an interpretation: user owns the interpretation or has delete access. The result of deleting an interpretation is that all the comments are deleted.
To delete a comment from an interpretation: user owns the comment, or user is a superuser.

Note the inconsistent behavior: the owner of the interpretation can delete a comment if they delete their whole interpretation. However he/she cannot delete individual comments.