Add snackbar that shows message about loading dashboard (#138)
If loading the dashboard takes more than 500ms, then show a snackbar message indicating that loading is in progress.

Note that action uses the loadingDashboardMsg object rather than creating a new object for each message. This is because of some buggy behaviour that is described in the MUI snackbar documentation (v0.20).