Use css height instead of display flex for plugin items when in view mode (#152)
Fix for the issue where vertical scrollbar appears on the plugin item after expanding then hiding the interpretations while also changing vis types. The display: flex was adjusting the height, but it should stay fixed while in view mode. When in edit mode, it should still use display:flex as the items might get dragged to a different size.