Use Trans component to display the styled delete confirm message (#165)
The dashboard name should be in bold to give it emphasis. In order to do this with internationalization, use the Trans component. The Trans component requires i18n in the context, so i18n was added at the root context (in App.js) so it will be available for any situations where Trans component is needed.