Respect user setting for control bar height when clicking on the CLOSE button (#168)
When the user clicks the CLOSE button in the dashboard controlbar, the height should be restored to the user's height setting, not to the minumum height. if the user setting is the same as the maximum height, then clicking on the CLOSE/VIEW_ALL_DASHBOARDS will have no effect, but the button text will change so the user knows that the action has taken place.