Integrate d2-i18n (#174)
Integrates d2-i18n library. d2-i18n documentation can be found at https://github.com/dhis2/d2-i18n/blob/master/README.md.

Note that the <strong> tag had to be removed around the delete dashboard name, and was replaced by double quotes instead. It is not possible to use the <Trans> component since it has problems with RTL languages. So then the choice was between providing a properly interpolated string without styling, or a non-interpolated string but keeping the styling. In the end we decided that replacing the <strong> with quotation marks was the best compromise.