Feature/edit interpretations [DHIS2-3430] (#191)
This feature enables editing of posted interpretations, and interpretation comments (replies) on analytic objects in the dashboard app.
The edit functionality are only available for the owner of the submitted interpretation.  

d2 version have been bumped from 29 to 30.0.2 in the package.json file, in order to dispatch plain text of updated interpretations through d2's .update() function.

What was added in this PR:

Interpretation.js
- this.state.editId (Keeping track of which interpretation, or interpretation comment that have pressed the edit button)
-  editButton
- toggleCommentField (earlier: showCommentField)
- submitComment (editing original interpretations, or interpretation comment)
- renderCommentOrEditField (Rendering InputField, or actual interpretation text, depending whether the edit button is pressed or not)
- toggleEdit (Assigning the id of the relevant interpretation, or interpretation comment to state variable editId)

- hasDeleteAccess (earlier: canDeleteInterpretation)
removed conditional check of evaluating whether access.delete is set to true or not.

actions.js (interpretations)
- tUpdateInterpretation (dispatching the action, updating the API endpoint, and the redux store)
- tUpdateInterpretationComment (dispatching the action, and update redux store)

interpretations.js (reducer)
- action type RECEIVED_INTERPRETATION (earlier: ADD_INTERPRETATION)
Responsible for both adding an interpretation, or edit an existing. 

Additionally, deleting interpretations are now only available for the owner, and super user (earlier, deleting interpretations were enabled if the access.delete property of the interpretation were set to true).