fix/no-access - Clearly separate between new and edit mode [DHIS2-4091] (#201)
Fix bug where user could not create a new dashboard, because the access check was based on the access level for the dashboard being viewed when the new button was clicked. The solution was to clearly separate new and edit mode when rendering the dashboard.

The fix involved a bit of refactoring to have similar components for View, Edit, and New states.