DHIS2-4749: fix comparison for finding app widget (#209)
It seems that by coincidence folderName was matching appKey. But
folderName has changed now, and in any case it was wrong to compare
folderName with appKey as key is what is actually read and written in
dashboardItems and comes from the same object present in both api/apps and api/dashboard/q responses.