Add reducer/actions for editItemFilters
This is used to differentiate between applied filters (itemFilters in
the store) and selected items in the filters, which is needed by the
dimension dialogs when the user interacts with the "Add filter" feature
in the app.
The Confirm button is what copies the selected filters in itemFilters
and that is what is used by the app for applying the filters to the
items.