feat: condense item header options into a single menu (#568)
All options in the dashboard item, for a visualization, are now moved to a single button with a dropdown menu. This includes the "View as" options for switching between Map, Table, Chart, as well as show/hideing the interpretations and open the Vis in the corresponding app.

This commit also fixes wrapping issues when for the item title when the item is narrow.