fix: apply scrollbar to the dashboard area below the control bars and headerbar [DHIS2-9371] (#1034)
The overflow property of app-shell-app has to be overridden to achieve the scrollbar on the dashboard area. Then, the height of the dashboard area is calculated from window.innerHeight - any visible control/header bars. A few margin and padding updates were needed, which affected the print "page" size.

As part of the implementation, the print actions bar is now fixed and stays in place when scrolling through the print dashboard.