fix: include title page in edit preview but scroll to below it [DHIS2-9417] (#1038)
In contrast to the Print Layout (from View mode), the Edit print preview did not have a title page, which means there was no page break prior to the first item page. This means the first item page had more vertical space in Edit print preview than in Print Layout. The result could lead to more items getting on the first page in Edit preview, than the actual print.

The solution is to include the title page in Edit preview as well, so that Edit preview and Print Layout are identical. Then, to allow the user to focus on the important stuff (item print layout), the Edit preview is auto scrolled to the top of the first item page.