fix: patch rgl with fix (#1046)
There is a bug in react-grid-layout that only occurs when there are static items in the layout, which there are when we use page breaks for the print view. For now, I've locked RGL to 1.1.0 so that dependabot doesn't bump the version and invalidate the patch.

This fix applies a patch to the react-grid-layout node module. The fix will be turned into a PR in the source repo.

What this patch does:

The function where this change is made is responsible for detecting collisions between items and increasing the y value on an item if it collides with one above it. It goes through the sorted item list (sorted by increasing y,x) and recursively compares the item with items below it. When there are no statics, then it can break out of the list when the item below it doesn't collide. But with the static items, a lot of the dashboard items are going to get pushed down and the code cannot assume that just because the item below didn't collide, that other items aren't going to get pushed down.