fix: filter bar position (DHIS2-9453)
The calculation of the position of the filter bar was based on the
height of the control bar.
After the change for the scrollbar, the filter bar was positioned lower
than where it should have.
Removing the height of the control bar from the top position of the
filter bar puts it back in the right place.