fix: changing View As on one item was incorrectly causing other items to also change view [DHIS2-9590] (#1111)
Fixes:

* if 2 dashboard items have the same visualization, switching the "View as" on one item should not affect the other item
* if 2 dashboards have the same visualization, and the "View as" has been switched on one dashboard, it should not affect the other dashboard
* also fixes [DHIS2-9589]: Edit mode should show the original visualization type, not the "view as" type.

Fix for 9590 involves separating the "Active type" from the "Visualization" in redux, since the visualization is in a map keyed by visualization id (which can be shared between dashboard items). Fix also includes clearing out some of the redux properties when switching dashboards (clear visualizations)

Fix for 9589 was to choose the vis type to display based on whether the user is currently editing. In edit mode (and edit print preview), it should show the original visualization