refactor: workaround the lack of d2.model for the TranslationsDialog
The TranslationsDialog expects a d2 model representing the object on
which apply translations.
If the object passed has the href attribute, the dialog does not try to
access the api endpoint via the modelDefinition.
These changes basically simulate a model by adding the modelDefinition
object and make the dialog api requests work by making sure the object has the href property.