feat: responsive - redirect to view when url of /edit or /new manually entered [DHIS2-10138] (#1473)
If the app is displayed in small screen, then edit and new are not supported. If the user manually enters the /edit or /new route, then the app will redirect back to the view mode of the same dashboard that was being viewed.

Note that if the user was already in /edit or /new, and then changed the screen to <=480px, then they won't get redirected, but a message will be displayed about not being able to edit in small screen.