fix: add small More button for small screen [DHIS2-10424] (#1511)
Issue was that the More button is together in a button strip containing other buttons. The other buttons are hidden in small screen, but More button is still available, albeit in "small". This conversion to "small" was being handled by javascript. So there was a lag before the button changed.

The solution is to have two More buttons and they are hidden or shown using media query.