fix: remove div around content that blocked scrolling while adding items to new dashboard (#1573)
Fix: remove div around new dashboard that blocked scrolling when new items were added

Refactor:

fixed lint error caused by the import of EditTitleBar which was also a const export (no more). As a result, the jest test needed to be updated to testing-library with a mock redux store. This resulted in several snapshot updates.