fix: preserve sharing and translations when saving dashboard (#1611)
After the switch from d2 to DataQuery, translations were being deleted when a dashboard is saved. This is because the skipTranslation=true param was missing. This and skipSharing=true have been added now, but the api currently has a bug in that it does not observe these params. The backend fix is here: dhis2/dhis2-core#7538

Changes to EditBar: there is no need to set the dashboard display name in the onTranslationsSaved callback, since the display name will be fetched with the updated dashboard when user returns to ViewMode. So code for setting display name in actions and reducers was removed.