fix: check dom element directly to determine if element is fullscreen in order to avoid unnecessary visualization rerenders (#1672)
Instead of setting a state to trigger changes related to going fullscreen, the components that need to observe the state of fullscreen will now use the native web api and dom elements. This avoids a second render that was happening when going fullscreen.

The onResizeStop listener isn't needed anymore since each plugin individually handles size changes (in MapPlugin, LegacyPlugin, DV Plugin)

Some of the files are just function renaming.

Other improvement: do not create an ItemContextMenu when in edit mode.