feat: add confirmation dialog for discarding changes to dashboard (#1713)
This change was requested as part of the Dashboard PWA [DHIS2-10874]. (when editing a dashboard and connectivity lost, then the dialog can be used to warn user of lost changes if they exit).

If a user makes any change (name, description, changes to items or filter settings), then the dirty state will be set and the dialog will be triggered when exiting without saving. However, adding translations does not affect the dirty status of the dashboard.