chore: move showDescription to root of redux structure (#1732)
showDescription has been stored within the redux property selected even though it has nothing to do with the selected dashboard (it is a user preference). Makes most sense to keep it at the root of the redux structure. This is also in preparation to needed changes to selected for the offline dashboard features