fix: save starred dashboard should not remove the star [DHIS2-11089] (#1757)
The fields favorite and favorites were missing in the save dashboard payload, resulting in the dashboard's "starred" status being set to false.

Also, created and lastUpdated are not used anywhere in the dashboard app so there is no point in prettifying their format (in getCustomDashboards). They are only needed for save, and even then, they probably aren't needed, but playing it safe here.

Cypress tests:
A cypress test was added to check the scenario outlined in the bug report.
Additionally, all the filter restrict tests were re-enabled after fixing to ensure the created dashboard is deleted after the last test.