fix: add onError callback to handle failed hart from DV plugin [DHIS2-11303] (#1832)
Add callback for errors from DataVisualizerPlugin. The implementation is based on the sketch from DHIS2-2839.

Changes to VisualizationItem/Item look like a lot, but the only change was to nest another

so that the error message could be centered vertically.
The key part of the fix is in DataVisualizerPlugin with the error handler, and resetting the error state when either the filter or the visualization type changes (i.e., View As Table).