fix: set correct item height when adding dashboard items in edit mode [v36] (#1869)
Backport of #1868

Remove the nested div that was causing the height for a visualization to not be set during edit mode. Fix was to restore the previous DOM structure, and just add a div around the error message with the needed height.

This bug was introduced by the fix for DHIS2-11303. PR that introduced the bug: #1832
The other change in this PR is to improve the styling of the error message.