refactor: move notice to components folder (#1908)
The Notice component will be used by ViewDashboard for upcoming offline implementation, so it needs to move to the shared components folder.

The cypress tests check this notice box, so if the tests pass, then the functionality should be 👍

No functional or visual changes. The notSupportedNotice css class did not exist so it was removed.