fix: adjust z-index so tooltip on Show more/fewer dashboards button shows (#1912)
https://jira.dhis2.org/browse/DHIS2-11647

Because of competing z-indexes on the component cover and tooltip, the tooltip was hidden when the dashboards bar was expanded.

Components should be stacked from bottom to top: cover, dashboards bar, tooltip.
Previously, the z-index of these was: 2000, 2001, 2000 which means the tooltip wasn't visible.
With the new code, the z-index is: 1998, 1999, 2000. (Cannot use ui ComponentCover since z-index is not configurable).