fix: add userId to cache section id so cached dashboards are user-specific (#1954)
Partial fix for https://jira.dhis2.org/browse/DHIS2-11759

The cache section id now includes the userId so that cached dashboards are specific to the user. One thing that cannot be fixed on the dashboard side is the caching of the list of dashboards, which is also user-specific. But with the implementation of cache clearing on manual logout, this is unlikely to come up as a problem.

Notes about the technical solution:
The useCacheableSection hook from app-runtime has been wrapped into a hook that sets the cacheSectionId with the userId-dashboardId. This was to reduce the number of places that needed to fetch the userId from d2.