fix: when loading a dashboard set the loaded flag after the fetch so state doesn't toggle back and forth [v37]
loaded was set to true before firing the fetch dashboards request. When that request failed due to the cache being hacked and deleted, then catch set loaded back to false. The result was that loaded kept toggling between true and false.

In addition, after discussion with Kai, it is not necessary to listen to recordingState to determine if loading is necessary. Note that removal of that recordingState doesn't impact any behavior.