fix: if map contains an ee layer, then do not display it when offline [v37] (#2041)
Updated fix for https://jira.dhis2.org/browse/DHIS2-11908

Instead of showing partial maps (e.g. maps with all its layers except ee layers) when offline, do not show the map at all if it has an EE layer. There was a lot of complexity and potential bugs with the previous solution, which was removing ee layers and displaying the rest of the layers.