feat: landing page
See https://jira.dhis2.org/browse/DHIS2-11989 for description/requirements.

Adds a new route: /start

When the app loads check if the new system setting https://jira.dhis2.org/browse/DHIS2-11993 is true, if yes navigate to /start.

Adds a "Close dashboard" menu item in the "More" menu on the dashboard toolbar. This navigates to /start.

Displays an available-offline indicator next to the most viewed dashboards.

Adds a "Go to start page" link to the not-available-offline notice box.