fix: org unit filter dialog crash (DHIS2-13816) (#2118)
Fixes https://dhis2.atlassian.net/browse/DHIS2-13816.

The OrgUnitDialog from analytics had a breaking change in v21.0.0, a different set of props is needed to make it work.

Several Cypress tests were updated to work with the new OrgUnitDimension component. There are still 4 cypress tests failing because of a real bug in the app.