feat: add custom message for when the plugin is not installed
More friendly than the generic error message.
In this case there isn't any problem with the dashboard item, but the
plugin for rendering the item is not installed.
We inform the user about it and suggest to install the missing app from
the app hub.

(cherry picked from commit d9a7f0bcfa2eff6d56314b609ec6496769a4e402)