feat: implement dashboard slideshow (#3081)
Implements https://dhis2.atlassian.net/browse/DHIS2-13038

Details:

1. When entering the slideshow/fullscreen, the parent div of the ItemGrid is put
into browser fullscreen. That same div remains in fullscreen the entire time, while
the dashboard items are displayed/hidden in turn using css opacity. Css opacity is
used instead of display or visibility so that the items can first be resized to
fullscreen while "invisible", and then be made visible. This means that the user
doesn't see a flash of the visualization in its regular item size before it gets resized.

2. Progressive loading. When the dashboard goes into fullscreen/slideshow, then progressive
loading will load the next and previous items in the fullscreen display order instead of
being based on proximity to the viewport.

3. RTL support has been implemented for the slideshow navigation buttons. It was decided
after discussions with in-house RTL experts that the order
of the dashboard items would remain LTR.