feat: maximize use of the available screen space by reducing whitespace in the dashboard item grid
Implements part of DHIS2-18441 (#3165)

Note that these changes are not included in the design doc

---

### Key features

1. Reduces whitespace in the dashboard item grid
2. Adjusts layout when a scrollbar appears

---

### Description

In this PR we adjust the overall grid styles a bit to allow dashboard-items to be a little bit
closer together. Additionally we wanted to solve the problem that our grid layout didn't take the
presence/absence of a scrollbar into account. Before we simply had reserved sufficient margin on the
right so that things would still look OK and function correctly when a scrollbar was present. In
this PR an additional container was added which does not scroll and thus will reduce in width when
its parent gets a scrollbar. Because it has a `ResizeObserver` attached to the DOM element we can
monitor when this happens and trigger a state change. We've done this using a Provider/Consumer and
a hook so that descendants can easily access this state.

---

### References
JIRA issue DHIS2-18441: https://dhis2.atlassian.net/browse/DHIS2-18441
Design doc: https://docs.google.com/document/d/1c8Ll1aLbFYwU8HYsyDlH1wk_QKTrUXwNanCiqayczOM