fix: not possible to see more or search for App items when editing dashboard (#3186)
Fixes DHIS2-18835

Fix: the Item selector list is populated based on the item type, but for some reason
the code for setting which types should "Show more" was using itemType.id, which APP lacked.
Now, it will consistently check the type rather than id. But also, the id property for APP
was added to the itemTypes object, for code consistency.