chore: add unit tests and implement feedback (#3206)
* test: add unit tests for superset embedded fields state hook

* fix: use regular radio inputs in choose-dashboard-type modal

* fix: apply cursor style to outer container of radio

* fix: ensure selected is an object in edit dashboard unit test

* fix: adjust embedded dashboard fields

Since toggling the filter visibility is not working via the SDK
we have removed that option. And we have replaced it with an
option to expand/collapse the filters instead.

* chore: remove unused import

* fix: adjust buttons texts to design specs

* fix: make choose-dashboard-type modal dismissable

* chore: upgrade reselect

* fix: use memoized selector for embedData to avoid unnecessary dashboard reloads

* refactor: use ms prefix for reselect memoized selectors

* fix: adjust import name

* fix: add padding below description for embedded dashboards

While the dashboards themselves have some whitespace above
the filter controls section does not, so some whitespace
below the description is needed to avoid it touching the
filter controls.
# Conflicts:
#	i18n/en.pot