chore: remove d2, and Event Charts/Reports auth header used in development mode (#3204)
The setting of the authentication header has caused a lot of headaches when testing
with non admin users as the app switches back to admin after login if an EC or ER
is in the dashboard.
This is most likely only something happening during in-house development, but nevertheless,
the potential for it to impact actual installations is real if they were to have
REACT_APP_DHIS2_AUTHORIZATION defined in their environment.

This header is no longer needed now that Event Charts and Event Reports will soon start
using withCredentials, which allows CORS requests.

Other changes:

All remaining uses of d2 are now migrated to app-platform dataEngine.
** The userDataStore which is used when setting the show/hide description
** The utility function generateUid was copied from d2 into this repo.
** The api request fetchVisualization has been migrated.
The show/hide description cypress test was reimplemented as a standard
non-cucumber test, as that is where we are headed. Thus, we need side-by-side
configurations for regular tests and cucumber tests. So, all the cucumber tests
were moved to a e2e_cucumber subfolder, and non-cucumber config was added.