fix: ensure app version is returned for bundled apps
For apps that are bundled and don't have a manually installed version,
api/apps doesn't return anything.
This causes dashboard to think there isn't an installed version of the
app/plugin and shows a warning in the dashboard item.
The app version should be also looked up in the bundled apps list.