feat: External Superset dashboards [DHIS2-17891] (#3205)
When the app starts it detects if the backend supports embedded Superset dashboards. If this is the
case, when creating a dashboard, the user will be presented with a choice to create an internal or
external dashboard. A new embedded Superset dashboard requires a Superset embed ID which is used to
show the external Superset dashboard in an iframe. Embedded dashboards can be starred, edited and
shared like internal dashboards, but they do not have a print view or slideshow mode.