fix: support bundled LL app (DHIS2-17400) (#3238)
Fixes https://dhis2.atlassian.net/browse/DHIS2-17400

Prepares the dashboard app to work with Line Listing app as a bundled app.

Since api/apps does not return bundled apps that don't have a manually installed version, a hardcoded path is needed for pointing the iframe to the correct plugin. This is the same approach we already use for DV and Maps. This change shouldn't affect the existing way of loading the LL plugin as currently the pluginLaunchUrl from api/apps is always used if LL app is installed. The manually installed versions always take precedence, the bundled apps hardcoded paths are used as fallback.