feat: allow widgets to be filtered by period [DHIS2-19186] (#3255)
Implements DHIS2-19186

Key features

Add support for adding pe (period) as a query parameter in the iframe source URL.
This update ensures that the pe (period) filter is included in the iframe URL for
dashboard items when the filter is applied. If the pe filter is present in itemFilters,
the periods are concatenated using , as the delimiter and added as a period
query parameter in the iframe source.